src/test/kotlin/chess/rules/BishopRulesTest.kt
package chess.rules

import chess.board.Board
import chess.rules.exceptions.EmptyTileException
import chess.rules.exceptions.WrongRuleException
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Test
import kotlin.test.assertFailsWith

class BishopRulesTest {
    private val bishopRules: BishopRules = BishopRules
    private var board: Board = Board()

    @BeforeEach
    fun setUp() {
        board = Board()
    }

    @Test
    fun whenGetWrongPiece_ThrowException() {
        assertFailsWith<WrongRuleException> { bishopRules.calculatePossibleMoves(2, 7,board) }
    }

    @Test
    fun whenGetEmptyTile_ThrowException() {
        assertFailsWith<EmptyTileException> { bishopRules.calculatePossibleMoves(3, 7,board) }
    }
}